+++ andino_firmware/andino_firmware.ino
@@ -1,0 +1,4 @@
+/*
+ * Arduino IDE sketch file. This file is required to be able to compile and
+ * upload the Andino firmware using Arduino IDE.
+ */
